automatically remove stale build label after the PR has been updated
