+++ constants.js
@@ -46,10 +46,11 @@ const issuesLabelCheck = 'issues-labeled-check';
 const issuesAssignedCheck = 'issues-assigned-check';
 const forcePushCheck = 'force-push-check';
 const pullRequestReviewCheck = 'pr-review-check';
-const codeOwnerCheck = 'code-owner-check'
+const codeOwnerCheck = 'code-owner-check';
 const ciFailureCheck = 'ci-failure-check';
 const updateWithDevelopCheck = 'update-with-develop-check';
 const respondToReviewCheck = 'respond-to-review-check';
+const oldBuildLabelCheck = 'old-build-label-check';
 
 const checksWhitelist = {
   'oppia-android': {
@@ -85,7 +86,8 @@ const checksWhitelist = {
       mergeConflictCheck,
       jobCheck,
       modelCheck,
-      codeOwnerCheck
+      codeOwnerCheck,
+      oldBuildLabelCheck
     ],
     [closeEvent]: [allMergeConflictCheck, updateWithDevelopCheck],
     [editEvent]: [wipCheck],
@@ -147,7 +149,8 @@ module.exports.pullRequestReviewCheck = pullRequestReviewCheck;
 module.exports.codeOwnerCheck = codeOwnerCheck;
 module.exports.ciFailureCheck = ciFailureCheck;
 module.exports.updateWithDevelopCheck = updateWithDevelopCheck;
-module.exports.respondToReviewCheck = respondToReviewCheck
+module.exports.respondToReviewCheck = respondToReviewCheck;
+module.exports.oldBuildLabelCheck = oldBuildLabelCheck;
 
 module.exports.getBlacklistedAuthors = function() {
   return blacklistedAuthors;
