+++ lib/utils.js
@@ -31,6 +31,7 @@ const SUCCESS_STATUS = 204;
 const THREE_MINUTES = 60 * 1000 * 3;
 let MIN_BUILD_DATE = new Date();
 MIN_BUILD_DATE.setDate(MIN_BUILD_DATE.getDate() - 2); // 2 days prior.
+const OLD_BUILD_LABEL = "PR: don't merge - STALE BUILD";
 
 /**
  * Gets all the changed files in the repo
@@ -419,6 +420,7 @@ module.exports = {
   THREE_MINUTES,
   LABELS_EXCLUDED_FROM_CODEOWNER_ASSIGNMENT,
   DATASTORE_LABEL,
+  OLD_BUILD_LABEL,
   getAllChangedFiles,
   getNameString,
   getNewItemsFromFileByRegex,
